+++ CHANGELOG.rst
@@ -2,4 +2,6 @@ Changelog
 =========
 
 
-- Add ``search`` function to search for a given pattern in the text provided.
+- Add ``search`` method to search for a given pattern in the text provided.
+- Add ``iterfiles` method to yield all the file paths in a given folder path.
+- Add ``is_executable`` method to validate whether the given file is a executable or not.